requirements.txt
# Core dependencies
requests>=2.28.0
paramiko>=2.11.0
beautifulsoup4>=4.11.0
urllib3>=1.26.0
cryptography>=3.4.0
reportlab>=3.6.0
python-nmap>=0.7.1
shodan>=1.28.0
colorama>=0.4.4
pytest>=7.1.0

# Optional dependencies (install manually if needed)
# PyQt5>=5.15.0  # For GUI (install separately)
# scapy>=2.4.5   # For advanced network scanning
dnspython>=2.2.0  # For DNS operations and subdomain enumeration
# selenium>=4.8.0  # For web automation
# sqlalchemy>=1.4.0  # For advanced database operations

